requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bitarray==2.8.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cytoolz==0.12.2
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
flipside==2.0.8
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.1.0
parsimonious==0.9.0
protobuf==4.24.1
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-telegram-bot==20.5
python3-openid==3.2.0
pytz==2023.3
pyunormalize==15.0.0
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.29.0
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.9.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
toolz==0.12.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
web3==6.10.0
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
